els-ASAP auto commit file/tag cmake-2.8.12.2-4.el6.src.rpm

+++ src/Source/cmDepends.cxx
@@ -1,19 +1,14 @@
-/*=========================================================================
+/*============================================================================
+  CMake - Cross Platform Makefile Generator
+  Copyright 2000-2009 Kitware, Inc., Insight Software Consortium
 
-  Program:   CMake - Cross-Platform Makefile Generator
-  Module:    $RCSfile: cmDepends.cxx,v $
-  Language:  C++
-  Date:      $Date: 2008-05-15 19:39:50 $
-  Version:   $Revision: 1.17.2.1 $
+  Distributed under the OSI-approved BSD License (the "License");
+  see accompanying file Copyright.txt for details.
 
-  Copyright (c) 2002 Kitware, Inc., Insight Consortium.  All rights reserved.
-  See Copyright.txt or http://www.cmake.org/HTML/Copyright.html for details.
-
-     This software is distributed WITHOUT ANY WARRANTY; without even
-     the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-     PURPOSE.  See the above copyright notices for more information.
-
-=========================================================================*/
+  This software is distributed WITHOUT ANY WARRANTY; without even the
+  implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+  See the License for more information.
+============================================================================*/
 #include "cmDepends.h"
 
 #include "cmLocalGenerator.h"
@@ -30,7 +25,7 @@ cmDepends::cmDepends(cmLocalGenerator* lg, const char* targetDir):
   Verbose(false),
   FileComparison(0),
   TargetDirectory(targetDir),
-  MaxPath(cmSystemTools::GetMaximumFilePathLength()),
+  MaxPath(16384),
   Dependee(new char[MaxPath]),
   Depender(new char[MaxPath])
 {
@@ -55,6 +50,7 @@ bool cmDepends::Write(std::ostream &makeDepends,
   std::vector<std::string> pairs;
   cmSystemTools::ExpandListArgument(srcStr, pairs);
 
+  std::map<std::string, std::set<std::string> > dependencies;
   for(std::vector<std::string>::iterator si = pairs.begin();
       si != pairs.end();)
     {
@@ -67,9 +63,14 @@ bool cmDepends::Write(std::ostream &makeDepends,
     obj = this->LocalGenerator->Convert(obj.c_str(),
                                         cmLocalGenerator::HOME_OUTPUT,
                                         cmLocalGenerator::MAKEFILE);
+    dependencies[obj].insert(src);
+    }
+  for(std::map<std::string, std::set<std::string> >::const_iterator
+      it = dependencies.begin(); it != dependencies.end(); ++it)
+    {
 
     // Write the dependencies for this pair.
-    if(!this->WriteDependencies(src.c_str(), obj.c_str(),
+    if(!this->WriteDependencies(it->second, it->first,
                                 makeDepends, internalDepends))
       {
       return false;
@@ -87,7 +88,8 @@ bool cmDepends::Finalize(std::ostream&,
 }
 
 //----------------------------------------------------------------------------
-bool cmDepends::Check(const char *makeFile, const char *internalFile)
+bool cmDepends::Check(const char *makeFile, const char *internalFile,
+                      std::map<std::string, DependencyVector>& validDeps)
 {
   // Dependency checks must be done in proper working directory.
   std::string oldcwd = ".";
@@ -102,7 +104,7 @@ bool cmDepends::Check(const char *makeFile, const char *internalFile)
   // Check whether dependencies must be regenerated.
   bool okay = true;
   std::ifstream fin(internalFile);
-  if(!(fin && this->CheckDependencies(fin)))
+  if(!(fin && this->CheckDependencies(fin, internalFile, validDeps)))
     {
     // Clear all dependencies so they will be regenerated.
     this->Clear(makeFile);
@@ -138,23 +140,29 @@ void cmDepends::Clear(const char *file)
 }
 
 //----------------------------------------------------------------------------
-bool cmDepends::WriteDependencies(const char*, const char*,
-                                  std::ostream&, std::ostream&)
+bool cmDepends::WriteDependencies(
+    const std::set<std::string>&, const std::string&,
+    std::ostream&, std::ostream&)
 {
   // This should be implemented by the subclass.
   return false;
 }
 
 //----------------------------------------------------------------------------
-bool cmDepends::CheckDependencies(std::istream& internalDepends)
+bool cmDepends::CheckDependencies(std::istream& internalDepends,
+                                  const char* internalDependsFileName,
+                            std::map<std::string, DependencyVector>& validDeps)
 {
   // Parse dependencies from the stream.  If any dependee is missing
   // or newer than the depender then dependencies should be
   // regenerated.
   bool okay = true;
+  bool dependerExists = false;
+  DependencyVector* currentDependencies = 0;
+
   while(internalDepends.getline(this->Dependee, this->MaxPath))
     {
-    if ( this->Dependee[0] == 0 || this->Dependee[0] == '#' || 
+    if ( this->Dependee[0] == 0 || this->Dependee[0] == '#' ||
          this->Dependee[0] == '\r' )
       {
       continue;
@@ -168,6 +176,16 @@ bool cmDepends::CheckDependencies(std::istream& internalDepends)
     if ( this->Dependee[0] != ' ' )
       {
       memcpy(this->Depender, this->Dependee, len+1);
+      // Calling FileExists() for the depender here saves in many cases 50%
+      // of the calls to FileExists() further down in the loop. E.g. for
+      // kdelibs/khtml this reduces the number of calls from 184k down to 92k,
+      // or the time for cmake -E cmake_depends from 0.3 s down to 0.21 s.
+      dependerExists = cmSystemTools::FileExists(this->Depender);
+      // If we erase validDeps[this->Depender] by overwriting it with an empty
+      // vector, we lose dependencies for dependers that have multiple
+      // entries. No need to initialize the entry, std::map will do so on first
+      // access.
+      currentDependencies = &validDeps[this->Depender];
       continue;
       }
     /*
@@ -178,11 +196,19 @@ bool cmDepends::CheckDependencies(std::istream& internalDepends)
       }
       */
 
-    // Dependencies must be regenerated if the dependee does not exist
-    // or if the depender exists and is older than the dependee.
+    // Dependencies must be regenerated
+    // * if the dependee does not exist
+    // * if the depender exists and is older than the dependee.
+    // * if the depender does not exist, but the dependee is newer than the
+    //   depends file
     bool regenerate = false;
     const char* dependee = this->Dependee+1;
     const char* depender = this->Depender;
+    if (currentDependencies != 0)
+      {
+      currentDependencies->push_back(dependee);
+      }
+
     if(!cmSystemTools::FileExists(dependee))
       {
       // The dependee does not exist.
@@ -198,24 +224,49 @@ bool cmDepends::CheckDependencies(std::istream& internalDepends)
         cmSystemTools::Stdout(msg.str().c_str());
         }
       }
-    else if(cmSystemTools::FileExists(depender))
+    else
       {
-      // The dependee and depender both exist.  Compare file times.
-      int result = 0;
-      if((!this->FileComparison->FileTimeCompare(depender, dependee,
-                                             &result) || result < 0))
+      if(dependerExists)
         {
-        // The depender is older than the dependee.
-        regenerate = true;
-
-        // Print verbose output.
-        if(this->Verbose)
+        // The dependee and depender both exist.  Compare file times.
+        int result = 0;
+        if((!this->FileComparison->FileTimeCompare(depender, dependee,
+                                              &result) || result < 0))
           {
-          cmOStringStream msg;
-          msg << "Dependee \"" << dependee
-              << "\" is newer than depender \""
-              << depender << "\"." << std::endl;
-          cmSystemTools::Stdout(msg.str().c_str());
+          // The depender is older than the dependee.
+          regenerate = true;
+
+          // Print verbose output.
+          if(this->Verbose)
+            {
+            cmOStringStream msg;
+            msg << "Dependee \"" << dependee
+                << "\" is newer than depender \""
+                << depender << "\"." << std::endl;
+            cmSystemTools::Stdout(msg.str().c_str());
+            }
+          }
+        }
+      else
+        {
+        // The dependee exists, but the depender doesn't. Regenerate if the
+        // internalDepends file is older than the dependee.
+        int result = 0;
+        if((!this->FileComparison->FileTimeCompare(internalDependsFileName,
+                                             dependee, &result) || result < 0))
+          {
+          // The depends-file is older than the dependee.
+          regenerate = true;
+
+          // Print verbose output.
+          if(this->Verbose)
+            {
+            cmOStringStream msg;
+            msg << "Dependee \"" << dependee
+                << "\" is newer than depends file \""
+                << internalDependsFileName << "\"." << std::endl;
+            cmSystemTools::Stdout(msg.str().c_str());
+            }
           }
         }
       }
@@ -224,8 +275,20 @@ bool cmDepends::CheckDependencies(std::istream& internalDepends)
       // Dependencies must be regenerated.
       okay = false;
 
+      // Remove the information of this depender from the map, it needs
+      // to be rescanned
+      if (currentDependencies != 0)
+        {
+        validDeps.erase(this->Depender);
+        currentDependencies = 0;
+        }
+
       // Remove the depender to be sure it is rebuilt.
-      cmSystemTools::RemoveFile(depender);
+      if (dependerExists)
+        {
+        cmSystemTools::RemoveFile(depender);
+        dependerExists = false;
+        }
       }
     }
 
@@ -235,12 +298,27 @@ bool cmDepends::CheckDependencies(std::istream& internalDepends)
 //----------------------------------------------------------------------------
 void cmDepends::SetIncludePathFromLanguage(const char* lang)
 {
+  // Look for the new per "TARGET_" variant first:
+  const char * includePath = 0;
   std::string includePathVar = "CMAKE_";
   includePathVar += lang;
-  includePathVar += "_INCLUDE_PATH";
+  includePathVar += "_TARGET_INCLUDE_PATH";
   cmMakefile* mf = this->LocalGenerator->GetMakefile();
-  if(const char* includePath = mf->GetDefinition(includePathVar.c_str()))
+  includePath = mf->GetDefinition(includePathVar.c_str());
+  if(includePath)
     {
     cmSystemTools::ExpandListArgument(includePath, this->IncludePath);
+    }
+  else
+    {
+    // Fallback to the old directory level variable if no per-target var:
+    includePathVar = "CMAKE_";
+    includePathVar += lang;
+    includePathVar += "_INCLUDE_PATH";
+    includePath = mf->GetDefinition(includePathVar.c_str());
+    if(includePath)
+      {
+      cmSystemTools::ExpandListArgument(includePath, this->IncludePath);
+      }
     }
 }
